Cams and holopads for cc (#375)
* Disambiguate existing holopad labels

* Add holopads multitudinous to CentComm

These are primarily for the AI's benefit, not for crew communication,
which is why there are so many of them. The point is to let the AI have
a presence - and hear what people are saying - in places where
conversations are likely to occur. I tried to match labels on holopads
to labels on surveillance cameras.

* Expand CentComm camera coverage

Gets rid of some obnoxious blind spots for the poor AI's sake.

---------

Co-authored-by: Dylan Craine <dylan.craine@deepwatercreations.com>