D1-Logic for GunGame (#454)
* Add the whole shebang

* Add more cover to various places

* Remove misleading electrical sign

That was there to match a sign in the original map, but the original map
doesn't have a chain-link fence that could be (but isn't) electrified
right there.

* Add Logic to the integration tests

Knew I was forgetting something.

Actually, that's not true. I forgot to know that I was forgetting
something.

* Fix whitespace

Bah, it looked fine on my machine!