Resomi QOL Update #02 - Let there be colours! (#512)
* move EVA helmet 1px to the left

* Give Resomi state to the sunhat

* Resomi Markings Updated With More Colours

* Re-write main Resomi.yml to add Creampie, added non-functional handcuff sprite.

* Add 11 new Resomi states for outerclothing and helmets.

* Move the tail render layer below the outersuit.

* Fix issue with new marking.

* Changed the randomized Resomi names.

Special thanks to everyone who helped filling them out.

* Fixed the other EVA helmet.

I almost forgot that there are two.

* Made the multi-coloured tail it's own marking.

* Resolved error with conflict merge.

* Made the updated resomi tail it's own marking.

* Bring back all deleted markings.

I don't want anyone's Resomis to be broken upon merge. Or for imports from other forks to not work. Its best if I just keep all the old ones in tact and not rip off the bandaid.

* Fix health and slow threshholds

Forgot this was missing. Resomi are back to being cringe.

* fixed error and "full" sprite

* Apply Centronias' changes P1

* Apply Centronias' changes P2

* Apply Centronias' changes P3

* Fix indentation

* Fix indentation

---------

Signed-off-by: Huaqas <zacswebb@gmail.com>