Forced Research Server & R&D Console To Require Access, Added Droppable Rocks from Asteroid Rock (#37)
# Research Server
Seeing a TODO list plopped in the ether that is the discord's development channel, I decided to contribute a literal cent's worth of time to fix something easy.

![image](/attachments/7612198a-fa84-47a0-ad9a-fc376454ac5f)

As per the todo list's demand: the research console no longer has public access: you need research access. If you'd like to use it, use a card that has access. I am not adding a changelog, this is an incredibly, pathetically small change that doesn't need a whole changelog attached to it.

# Droppable Rocks
In reference to [this](https://git.tc14.space/tc14/trailblazer-colony-14/issues/17) issue, I added droppable rocks from asteroid rock. It's within the asteroid.yml, rocks.yml and ore.yml files. There was little reason to create new files.

Currently asteroid rocks have a 10% chance of dropping something, which has a random [0,2] drop limit. The randomness of the actual stack drop varies. It had not crashed in testing, but MAY crash if the game expects the rocks to stack (they currently do not).

Co-authored-by: LukeZurg22 <11780262+LukeZurg22@users.noreply.github.com>
Reviewed-on: https://git.tc14.space/tc14/trailblazer-colony-14/pulls/37
Co-authored-by: LukeZurg22 <lukezurg22@noreply.localhost>
Co-committed-by: LukeZurg22 <lukezurg22@noreply.localhost>