Added Playing Cards and Nuno [ Estacao Pirata Port ] (#753)
* Port Playing Cards

Taken from MIT branch https://github.com/Day-OS/estacao-pirata-14/tree/cards

* Fix spelling error in rsi

* Fix error.

Shuffles some code around and deleted radius and it worked. Haven't tested the radial menus, so this could have completely broken it. We will see...

* Collapse the rsi

This is how they do it and I hate that it has to be this way

* Revert "Collapse the rsi"

This reverts commit c31dc00092c118d6142257b298332e756b84040c.

* Organize yaml and add Syndicate/Nanotrasen decks

* Add FTL Designation to Syndicate/Nanotrasen cards

* Swap from abstract to hide in spawn menu

* Re-re-collapse RSI

Unfortunately with the way its coded, it tried to look at the same directory for everything. I wanted everything to be nice and neat but the code does not support it.

* Give the cards a spawn location.

Don't forget to add playing cards to the game idiot!

* Remove disclaimer.

People should know how licensing works if they are PRing stuff.

* Replace noSpawn with HideSpawnMenu

Obsolete code replaced with based giga-chad code.

* Comment out Log code.

* Add abstract to base protos

* Neaten code and resolve fatal error.

Fatal error was the _audio.PlayPredicted line trying to play the pickup sound at the location of the deck. When you take the last card that deletes the deck and it tried to find the coordinates of a entity that doesn't exist. Causing fatal crash. I changed the location from the card deck to the player. This should solve the issue and since the player has to be near the deck to interact with it, it shouldn't be issue.

Can't abstract some of these entities. Thought I could, but it doesn't like it.

I also merged some lines that I'm pretty sure could be merged. Let me know if I fucked up cause I can't always tell.

* Replace Obsolete Code with Working Code

* Stopped one of the tests from yapping.

* Nitpick cleanup

* 4-space

still hate this file

* rewrite stuff

* smoll cleanup + fix whitespace

* fixes, etc.

* Single cards hands are no longer invisible (and shouldn't ever exist anyway)

* - Drawing cards into an empty hand from a deck maintains its flippedness
- Taking the second to last card out of a hand leaves the last last card in the same container the hand was in
- Cards can be flipped in-world by hovering and pressing E (or interact)
- Decks no longer have a little flange of sprite hanging off their sides

* per-entity strip menu obscuring

* Cards utilize strip menu obscuring

* Sprite Int Test

Most of the sprites in the card system are managed by custom components, so an empty png was added to stop the integration check from crying.

* Specify Back Sprites

* Split the .rsi.... Again....

* Sprite fixes.

* Remove unused sprites.

* Add Nuno™

Currently bugged, as the layered sprite approach for Nuno™ cards currently doesn't work with current sprite system.

* Fix Sprite Inconsistencies with Card

Bunch of random cards had different colours for the border on their face up side. This has been fixed.

* Fix Nuno's Sprite Layering, Tweak Nuno Sprites

Changed the way the layering is implemented in yaml so it actually works.
Made the playing cards 1 pixel taller. Why? Don't ask.
Modified all numbers for clarity and to be centered in the card.
Created a new Reverse sprite because the old one kinda sucks.
Changed the block card to skip. (It is called skip in the yaml).
Remove the border sprite and merge it into the background.

* Revert "Fix Nuno's Sprite Layering, Tweak Nuno Sprites"

This reverts commit 440bf38e49c031d362eea93ba854cbe57e525515.

* Add back Cent's fixes after i graciously deleted them.

* Fix Nuno's Sprite Layering, Tweak Nuno Sprites

Overwrite Cent's shit. But we good now.

* True

* Swap card sprites to layers.

Created layers from all the original cards.
Making custom decks will be much easier in the future.
I also fixed a visual error with box open/close sprites.
Also 4-space.

* Update base.yml

---------

Co-authored-by: Centronias <me@centronias.com>