Skills System: Part 1 [WIP] (#51)
This is still WIP: it's not yet playtest-ready, but it is close enough for me to closely monitor integration test results.
- [X] Slight UI fixes, actually load passions from DB into the character creation UI
- [x] Roundstart skill points distribution system based on passions
- [x] Basic implementation for at least 3 skills
- [x] A few admin commands

## What was changed in this PR? What goals are you trying to achieve with these changes?
Added a basic skills system: you can see your skills in the character menu, and set your passions in the character editor menu.
## Please provide technical details to the implementation of this feature, if it's complex enough.
SkillsPicker is the UI element responsible for setting passions in the character editor.
Content.Shared/_tc14/Skills contains most of the actual in-simulation logic.
There are also a lot of tweaks in upstream code, because we need to send skill data as part of CharacterData, as well as making changes to the UI.
Database stuff. Scary. We store pairs of SkillProtoId - Value with ProfileID.

## Please attach the media showcasing the feature.
TODO
## Requirements
- [X] I have read and am following the [Pull Request and Changelog Guidelines](https://docs.spacestation14.com/en/general-development/codebase-info/pull-request-guidelines.html).
- [X] I have read and am following the [TC14 Conventions](https://wiki.tc14.space/Conventions)

**Changelog**
I will not fill this out, we don't even have changelog publisher so far.

Co-authored-by: mjarduk <marduk@pissmail.com>
Reviewed-on: https://git.tc14.space/tc14/trailblazer-colony-14/pulls/51