Small cleanup of SoilManager (#80)
SoilManager was using some hardcoding - I made it slightly better; boyscout rule and so on.
I also changed some raw strings in ContentTileDefinition to EntProtoIds and added a simple boolean for chiseling the tile to indicate if it can be stripped to stone; this will be needed for the winter map variant.
Finally, fixed missing locstrings for tiles - no idea why linter wasn't picking it up.

Reviewed-on: https://git.tc14.space/tc14/trailblazer-colony-14/pulls/80
Co-authored-by: murphyneko <killerqueen1777@gmail.com>
Co-committed-by: murphyneko <killerqueen1777@gmail.com>